rbd: pass flags to rbd_req_sync_exec()
In order to allow both read request and write requests to be
initiated using rbd_req_sync_exec(), add an OSD flags value
which can be passed down to rbd_req_sync_op().  Rename the "data"
and "len" parameters to be more clear that they represent data
that is outbound.

At this point, this function is still only used (and works) for
write requests.

Signed-off-by: Alex Elder <elder@inktank.com>