rbd: use rbd_dev consistently
Most variables that represent a (struct rbd_device) are named
"rbd_dev", but in some cases "dev" is used instead.  Change all
the "dev" references so they use "rbd_dev" consistently, to make it
clear from the name that we're working with an RBD device (as
opposed to, for example, a (struct device)).  Similarly, change
the name of the "dev" field in (struct rbd_notify_info) to be
"rbd_dev".

Signed-off-by: Alex Elder <elder@inktank.com>